Refactor main.dart into widgets and pages
Split the monolithic `lib/main.dart` file into a modular structure:
- Extracted `ClickableLink` to `lib/widgets/clickable_link.dart`.
- Extracted `SettingsPage` to `lib/pages/settings_page.dart`.
- Extracted `MyHomePage` to `lib/pages/home_page.dart`.
- Cleaned up `lib/main.dart` to only contain the app entry point.

This improves maintainability and readability of the codebase.
